requirements.txt
tensorflow
PyOpenGL
PyOpenGL_accelerate
numpy